fix recursive stack overflow

+++ lua/src/a.h
@@ -139,7 +139,7 @@ void lua_array_num_new(lua_State *L, LUA_NUM const *ptr, unsigned int num);
 void lua_array_num_get(lua_State *L, int idx, LUA_NUM *ptr, unsigned int num);
 void lua_array_num_set(lua_State *L, int idx, LUA_NUM const *ptr, unsigned int num);
 
-LUA_NUM *lua_table_num_get(lua_State *L, int idx, LUA_NUM const *ptr, size_t *num);
+LUA_NUM *lua_table_num_get(lua_State *L, int idx, LUA_NUM const *ptr, unsigned int *num);
 
 void lua_stack_view(lua_State *L, unsigned int line);
 // #define LUA_STACK(L) lua_stack_view(L, __LINE__)
